requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-groq==0.1.0
langchain-huggingface==0.0.1
chromadb==0.4.24
pypdf==4.0.1
python-dotenv==1.0.0
huggingface-hub>=0.23.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai>=0.3.0

streamlit>=1.31.0
docx2txt>=0.8
langchain-google-genai>=1.0.0
